fixed time units not going down to 0, sticking to 5, even when 5 is the tile move cost

+++ Source/Architecture/Constants.hpp
@@ -26,5 +26,5 @@ struct MoveData
 {
 	float x = 0;
 	float y = 0;
-	float time_units = 0;
+	int time_units = 0;
 };
